requirements.txt
alabaster==1.0.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
babel==2.16.0
boto3==1.35.57
botocore==1.35.57
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
Collectfasta==3.2.0
colorama==0.4.6
coverage==7.6.4
crispy-bootstrap5==2024.10
cryptography==43.0.3
cssbeautifier==1.15.1
decorator==5.1.1
distlib==0.3.9
Django==5.0.9
django-allauth==65.1.0
django-anymail==12.0
django-cors-headers==4.6.0
django-coverage-plugin==3.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==5.0.0
django-nvd3==0.10.1
django-redis==5.4.0
django-storages==1.14.4
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djlint==1.36.1
docutils==0.21.2
EditorConfig==0.12.4
exceptiongroup==1.2.2
executing==2.1.0
factory_boy==3.3.1
Faker==30.8.2
fido2==1.1.3
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
identify==2.6.2
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.27
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy==1.13.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-c==3.2.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pytest==8.3.3
pytest-django==4.9.0
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
regex==2024.11.6
requests==2.32.3
ruff==0.7.2
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-autobuild==2024.10.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.2
termcolor==2.5.0
text-unidecode==1.3
tomli==2.0.2
tqdm==4.67.0
traitlets==5.14.3
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.0
Werkzeug==3.0.6
